I2C: Fix implementation (working DDC over I2C)
Numerous fixes and changes to the I2C implementation so that it actually
works correctly.  The reading and writing is still a little bit sketchy,
as we can't push more than 12 bytes at a time, and might need to flag
an ACK or so.  Also, abort situations might need a better exit.